feat(index): aggregate pushdown — parser, exec, prefilter wiring, tests
Lands the plan-time and execute-time halves of aggregate pushdown.
Not yet wired into the scanner.

Plan side (rust/lance-index/src/expression/):
- Moves scalar/expression.rs to expression/scalar.rs, paralleling the
  new expression/aggregate.rs.
- AnyAggregateQuery and AggregateQueryParser traits.
- AggregateIndexSearch leaf with optional index_name, parsed query,
  optional per-aggregate filter, and the original SELECT expression.
- CountQuery (basic / approx / distinct / approx_distinct) and
  CountQueryParser.

Scalar index trait (rust/lance-index/src/scalar.rs):
- ScalarIndex::calculate_aggregate returning a partial-state ArrowScalar;
  default-error stubs added to btree, bitmap, bloomfilter, inverted,
  json, label_list, ngram, rtree, zonemap, and the LogicalScalarIndex
  wrapper.
- Re-exports lance_arrow_scalar::ArrowScalar through scalar::.

Execute side (rust/lance/src/io/exec/aggregate_index.rs):
- AggregateIndexSearchExec emits one partial-state RecordBatch whose
  schema is the concatenation of state_fields() for each paired
  AggregateFunctionExpr, so a downstream AggregateExec(Final) consumes
  it unchanged.
- One optional child input — a ScalarIndexExec — supplies a prefilter
  RowAddrMask. The prefilter load and per-aggregate index loads run in
  parallel.
- Intersects fragment bitmaps across indexed aggregates, materializes
  the allow list as concrete [0..physical_rows) ranges (avoids the
  RoaringBitmap::full() inflation in RowAddrTreeMap::Sub), then composes
  prefilter ∩ fragments_allow − deletion_mask into a single AllowList.
- Calls calculate_aggregate per indexed aggregate; falls back to
  counting the combined mask directly when an aggregate is a
  non-distinct COUNT without an associated index.
- Unit tests cover try_new validation, the Full+Partial count helper,
  and end-to-end execution with no prefilter, an AllowList prefilter, a
  BlockList prefilter, and deletions.

Also includes aggregate-pushdown-research.md surveying how mature query
engines structure aggregate pushdown.
