+++ rust/lance-index/src/expression.rs
@@ -1,0 +1,9 @@
+//! Plan-time expression parsing for scalar and aggregate index pushdown.
+//!
+//! Both halves split a user expression into an index-evaluable leaf plus the
+//! residual computation: [`scalar`] parses `WHERE` clauses, [`aggregate`]
+//! parses `SELECT`-list aggregates. The execute-time consumers live under
+//! `lance::io::exec::{scalar_index, aggregate_index}`.
+
+pub mod aggregate;
+pub mod scalar;
