+++ rust/lance-index/src/lib.rs
@@ -20,6 +20,7 @@ use roaring::RoaringBitmap;
 use serde::{Deserialize, Serialize};
 use std::convert::TryFrom;
 
+pub mod expression;
 pub mod frag_reuse;
 pub mod mem_wal;
 pub mod metrics;
