+++ rust/lance-index/src/scalar/bitmap.rs
@@ -19,6 +19,7 @@ use datafusion::physical_plan::SendableRecordBatchStream;
 use datafusion_common::ScalarValue;
 use deepsize::DeepSizeOf;
 use futures::{StreamExt, TryStreamExt, stream};
+use lance_arrow_scalar::ArrowScalar;
 use lance_core::utils::mask::RowSetOps;
 use lance_core::{
     Error, ROW_ID, Result,
@@ -40,7 +41,9 @@ use super::{
     BuiltinIndexType, SargableQuery, ScalarIndexParams, SearchResult, btree::OrderableScalarValue,
 };
 use crate::pbold;
-use crate::{Index, IndexType, metrics::MetricsCollector};
+use crate::{
+    Index, IndexType, expression::aggregate::AnyAggregateQuery, metrics::MetricsCollector,
+};
 use crate::{
     frag_reuse::FragReuseIndex,
     progress::IndexBuildProgress,
@@ -605,6 +608,18 @@ impl ScalarIndex for BitmapIndex {
 
         let selection = NullableRowAddrSet::new(row_ids, null_row_ids.unwrap_or_default());
         Ok(SearchResult::Exact(selection))
+    }
+
+    async fn calculate_aggregate(
+        &self,
+        query: &dyn AnyAggregateQuery,
+        _filter: Option<SearchResult>,
+        _total_rows: u64,
+        _metrics: &dyn MetricsCollector,
+    ) -> Result<ArrowScalar> {
+        Err(Error::invalid_input(format!(
+            "this index cannot accelerate the aggregate {query:?}"
+        )))
     }
 
     fn can_remap(&self) -> bool {
